requirements.txt
google-api-python-client==2.108.0
pandas==2.1.3
python-dotenv==1.0.0
rich==13.7.0
streamlit==1.29.0
openai>=1.6.1 